+++ hs/src/HelVM/HelMA/Automaton/Instruction/Extras/PrintAsm.hs
@@ -1,0 +1,4 @@
+module HelVM.HelMA.Automaton.Instruction.Extras.PrintAsm where
+
+class PrintAsm a where
+  printAsm :: a -> Text