+++ hs/src/HelVM/HelMA/Automaton/Instruction/Extras/TextExtra.hs
@@ -1,0 +1,6 @@
+module HelVM.HelMA.Automaton.Instruction.Extras.TextExtra where
+
+import qualified Data.Text as Text
+
+toLowerShow :: Show i => i -> Text
+toLowerShow = Text.toLower . show