+++ hs/test/HelVM/HelMA/Automaton/OptimizerSpec.hs
@@ -1,10 +1,10 @@
 module HelVM.HelMA.Automaton.OptimizerSpec (spec) where
 
 import           HelVM.HelMA.Automaton.API.OptimizationLevel
-import           HelVM.HelMA.Automaton.Instruction
+import           HelVM.HelMA.Automaton.Instruction.Extras.Constructors
 import           HelVM.HelMA.Automaton.Optimizer
 
-import           Test.Hspec                                  (Spec, describe, it)
+import           Test.Hspec                                            (Spec, describe, it)
 import           Test.Hspec.Expectations.Pretty
 
 spec :: Spec