+++ docs/developers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # 📅 Revision history for HelMA
 
+## 0.8.4.7 -- 2023-04-16
+
+* Add optimize
+* Add pattern synonyms
+* Add pretty printer to ASM
+
 ## 0.8.4.6 -- 2023-04-10
 
 * Add optimize